requirements.txt
asgiref==3.7.2
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
distlib==0.3.8
Django==5.0.1
djangorestframework==3.16.0
Faker==20.0.0
faker-biology==0.6.2
faker-python==0.0.1
filelock==3.13.1
idna==3.10
pdfminer.six==20231228
platformdirs==4.1.0
protobuf==3.20.3
psycopg2==2.9.9
pycparser==2.22
pyodbc==5.0.1
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.8.2
requests==2.32.3
shiboken6==6.7.2
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tqdm==4.66.5
tzdata==2023.4
urllib3==2.3.0
virtualenv==20.25.0
